Allow "site" module to specify alias domains
There is a bit of assumption baked into this module about the fact that
the site is accessible at a URL in the format of
beta.aws.openoakland.org (i.e. a Route53 zone). By adding an "aliases"
option, we can tell Cloudfront that our site should also be accessible
via other domains, namely our preferred "openoakland.org".

The caveat here is that for the TLS certificate, Amazon makes us
validate all aliases. We can automatically validate the primary fqdn
since it's in Route53, but not any aliases. So, when running the
terraform for the first time, the user will have to manually create the
other domains (in Namecheap or wherever else) during the validation
step.